+++ headers/captura.h
@@ -1,7 +1,7 @@
 #ifndef CAPTURA_H //if not defined (si no esta definido)
 #define CAPTURA_H //se define la libreria para que no se duplique 
 
-extern int calificaciones[40][5];
+extern float calificaciones[40][5];
 extern char nombresEstudiantes[40][50];
 extern float promediosEstudiantes[40];
