+++ headers/modificar.h
@@ -1,7 +1,7 @@
 #ifndef MODIFICAR_H
 #define MODIFICAR_H
 // Declaraciones extern de todas las variables que este módulo necesita
-extern int calificaciones[40][5];
+extern float calificaciones[40][5];
 extern char nombresEstudiantes[40][50];
 extern float promediosEstudiantes[40];
 extern int numeroEvaluaciones;
